+++ docs/reference/changelog.md
@@ -8,6 +8,8 @@
   Added a `Suspend` control which suspends the Counterweight application while a user-supplied callback function runs.
   Counterweight will stop controlling the terminal while the callback runs, and will resume when the callback returns.
   This can be used to run a subprocess that also wants control of the terminal (e.g., a text editor).
+- [#88](https://github.com/JoshKarpel/counterweight/pull/88)
+  Implemented absolute positioning, which allows precisely positioning elements on the screen.
 
 ## `0.0.4`